requirements.txt
google-genai           # For generating the quote text (Gemini - New SDK)
gTTS                   # Google Text-to-Speech (Free)
edge-tts               # Microsoft Edge TTS (Natural voices for better quality)
moviepy==1.0.3         # Video editing (Stick to v1.0.3 for stability)
instagrapi             # Instagram Uploading & Session Management
Pillow                 # Image processing
python-dotenv          # To hide your passwords
imageio-ffmpeg         # Required for video rendering
pydub  
numpy
edge-tts